Fix submodule importing
Basically this patch fix the issue where no .js was actually being
called.

It also fix the _init() call by creating a new JSContext for each
submodule file (javascript file) and using this file as the module.

The import properties are copied from the original global Object to the
new one.